* dev:
  Fix the financial verification test.
  Bootstrapping/Shutdown: trading liquidity should shrink quickly and not be subject to the liquidity growth factor
  Lower the funding limit of test pool #7
  Bootstrapping/Shutdown: pool funding limit shouldn't be taken into account when defining the new target BNT trading liquidity
  .gitignore scripts/test-network.ts
  Remove test network script
  Revert README.md
  Improve documentation.
  Remove "Staking" from "StakingRewards"
  Remove "Staking" from "StakingRewards"
  Remove "Staking" from "StakingRewards"
  Adjust the `PoolLiquidityV1` structure with the `PoolLiquidity` structure.
  Ensure that pending rewards aren't being accrued for terminated programs
  Ensure to remove an inactive program, with no remaining rewards, from providers' program list only if their stakes are zero as well
  Replace `MathEx.toUint128` with OZ's `SafeCast.toUint128`.
  Move global function `toUint128` into library `MathEx`.
  Change the following storage fields from 256 bits to 128 bits: - `PoolLiquidity.bntTradingLiquidity` - `PoolLiquidity.baseTokenTradingLiquidity`

+++ contracts/pools/PoolMigrator.sol
@@ -19,8 +19,8 @@ interface IPoolCollectionBase {
 
 interface IPoolCollectionV1 is IPoolCollectionBase {
     struct PoolLiquidityV1 {
-        uint256 bntTradingLiquidity; // the BNT trading liquidity
-        uint256 baseTokenTradingLiquidity; // the base token trading liquidity
+        uint128 bntTradingLiquidity; // the BNT trading liquidity
+        uint128 baseTokenTradingLiquidity; // the base token trading liquidity
         uint256 stakedBalance; // the staked balance
     }
